Review fixes, split out test to own class
